Fields validation for model

+++ src/main/java/servingwebcontent/repository/DictionaryCounterpartyRepository.java
@@ -6,5 +6,5 @@ import servingwebcontent.entity.DictionaryCounterparty;
 import java.util.List;
 
 public interface DictionaryCounterpartyRepository extends JpaRepository<DictionaryCounterparty, Long> {
-    List<DictionaryCounterparty> findByNameOrInn(String name, Integer inn);
+    List<DictionaryCounterparty> findByNameOrInn(String name, Long inn);
 }
